(fix) CircleCI issues fixed

+++ src/main/java/com/greenfox/tribes1/Building/Building.java
@@ -15,7 +15,7 @@ public abstract class Building {
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   private Long id;
   private Long level;
-  private Double HP;
+  private Float HP;
   private Timestamp started_at;
   private Timestamp finished_at;
