+++ src/main/resources/db/migration/mysql/V1_2__init.sql
@@ -2,7 +2,7 @@ CREATE TABLE `building`
 (
   `id`          bigint(20) NOT NULL AUTO_INCREMENT,
   `level`       bigint(20),
-  `HP`          bigint(20),
+  `HP`          float (20),
   `started_at`  timestamp(6),
   `finished_at` timestamp(6),
   PRIMARY KEY (`id`)
